server/requirements.txt
boto3==1.33.12
botocore==1.33.12
Django==4.2.6
django_rest_knox==4.2.0
django_rest_passwordreset==1.4.1
djangorestframework==3.14.0
python-decouple==3.8
Requests==2.32.3
